[LayoutNG] OOFs and column balancing
OOFs should affect column balancing:
https://github.com/w3c/csswg-drafts/issues/6279.

As an initial implementation, this CL allows OOFs to affect the initial
column balancing pass, which won't handle every case, but it should
handle the cases of most interest.

To accomplish this, lay out the OOFs normally inside their containing
block during the initial column balancing pass, allowing the OOF to
affect the initial column block size estimate. In order for the OOF to
also affect the TallestUnbreakableBlockSize(), we needed to set up
the constraint space in this case for fragmentation and propagate
this information to the containing block's builder.

Because OOFs now affect the initial column block size, there ended up
being strange results when an OOF was discovered before a spanner. The
OOF would affect the column block size for the columns before the
spanner, but we would wait to layout the OOF officially until the
multicol had finished layout. Because of this, we ended up adding new
columns after the spanner with a smaller column block size.

To fix this, this CL adds logic to the column layout algorithm to lay
out any OOF descendants when we find a spanner using the newly added
NGOutOfFlowLayoutPart::HandleFragmentation() function. This ensures that
any OOF descendants found before a spanner are actually laid out before
the spanner and take up the block size that they had initially
contributed to. (See the updated unit tests as an example). Note that
this doesn't work for nested fragmentation, and the related unit tests
still result in slightly different behavior than the unnested case.

Bug: 1206618
Change-Id: If21e1cac970289e569be9ab1a79030329732d613